+++ CHANGELOG.md
@@ -3,5 +3,11 @@
 
 When stuff changes, it's described here.
 
+## 2015-04-02
+ * Fixed ga analytics id in config
+ * Use window._gaq instead of internal 
+ * Added product widget impression metrics
+ * Changed impression/interaction context ps to pw
+
 ## 2015-03-18
  * Initial import / setup